make BrowseResponse a DTO and move to root namespace
